split out telephone utils & SMSauth methods
