--- wsgi/normalizeTel.py
@@ -1,3 +1,0 @@
-def normalizeTel(tel):
-    nTel = re.sub(r'[^a-zA-Z0-9\+]','', tel)
-    return nTel
